updating editor
